Renamed HomeScreen.js to Welcome.js and corrected reference.
